pyproject.toml
[project]
name = "video-editing-api"
version = "0.1.0"
description = "FastAPI backend for creative video editing with multi-agent collaboration"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Your Team", email = "team@example.com"}
]

dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy>=2.0",
    "alembic>=1.12.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "aiofiles>=23.0",
    "httpx>=0.24.0",
    "pyyaml>=6.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]

[tool.uv]
index-url = "https://pypi.org/simple"

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
